Forward some audio-related keystrokes by default
Bluetooth headphones send `AudioPause` when tapped. If cursor is sitting
on the wrong screen, it won't toggle the audio on the correct machine.
So, always forward it (and `AudioPlay`) to the main box.

Support for `AudioPause` currently requires my fork of Synergy:

benizi/synergy@4ab2c1de873797310326521886873ca18ff8d623