Add restart logic into `serveConnection`
Lots of related refactoring.

Remove some uses of `check(error)` (which is a holdout from me hating
`error`), and instead return the error to the caller.

Use a `sync.WaitGroup` in `main()` instead of in `runRemoteLoop`.

Disclaimer: not sure this revision would compile. I'm splitting a bunch
of changes that have been sitting around for about 8 months into
mostly-logical groupings.